# buffer: fix DoS vector in atob