@chjj 2024-02-06
I didn't write a benchmark for the node.js codebase, but here are the benchmarks from [feross/buffer][1]:

Without this PR:

```
$ node perf/write-base64.js
BrowserBuffer#write(8192, "base64") x 2,770 ops/sec ±0.64% (85 runs sampled)
BrowserBuffer#toString(8192, "base64") x 17,933 ops/sec ±1.87% (89 runs sampled)
```

With this PR:

```
$ ~/node/out/Release/node perf/write-base64.js
BrowserBuffer#write(8192, "base64") x 37,795 ops/sec ±0.57% (93 runs sampled)
BrowserBuffer#toString(8192, "base64") x 21,842 ops/sec ±0.63% (93 runs sampled)
```

In other words, with this PR, `atob` is __13.6 times faster__ and `btoa` is 1.2 times faster.

Also, without this PR, the performance of `atob` is completely unpredictable. There could be a worst-case base64 string which triggers 70 iterations on the linear search (`ArrayPrototypeIndexOf`) on every character, which means the new implementation of the loop is actually __70 times__ faster in the best case. It's just 13.6 times faster in the common case (with randomized data).