@Uzlopak 2024-02-06
@nodejs/performance 

Should there be benchmarks? Maybe we should still consider the perf for btoa/atob.