@benjamingr 2024-02-06
I am fine with landing this (since it's an improvement) but no benchmarks.

(as a side note nice seeing you back!)