@lemire 2024-02-06
It would be useful to have some clarification.

@addaleax stated that these functions **should not be used**, and should therefore not be optimized.

It seems that this PR attempts to reverse an earlier decision. Am I correct?

(I am not expressing any opinion, I just want to understand.)