@chjj 2024-02-06
> @addaleax stated that these functions should not be used, and should therefore not be optimized.

I really don't understand this logic. If it's never supposed to be used, why implement `atob` at all then? It's a liability to have it be __70 times__ slower than it needs to be. You might as well remove it at that point. 

If it's truly the case that we're not allowed to optimize this function, then I will submit another PR which removes it, because it's irresponsible to have it in node.js given its performance issues (I still maintain that it is indeed a DoS vector).

As far as I know, `atob` is the only portable way to synchronously parse base64 in a web browser. Over the past several years, I've watched node.js go to great lengths to replicate web APIs so programmers have the ability to write code that runs in both node and the browser. I ask you:

- Is `EventTarget` unoptimized in favor of `EventEmitter`? 
- Is the Web Crypto API unoptimized in favor of `require('crypto')`?
- Is `TextEncoder` unoptimized in favor of `string_decoder`?

The answer to all of this is no. So why is `atob` unoptimized in favor of `Buffer.from(str, 'base64')`?

I understand that `Buffer` is a saner way to decode base64, but the web APIs don't offer any synchronous options aside from `atob`. It's certainly an unnecessary function for node.js to have, but note that [MDN does not][1] consider `atob` "legacy" or deprecated in any way. It's a standardized web API.

I suppose in the browser we could always just use [feross/buffer][2], but I would like to remind you that you are speaking to someone who is currently helping develop that library. In fact, it's the very reason I made this PR!

And for reference:

Here is [deno's implementation of atob][3]. It's calling out to an SIMD-optimized base64 decoding function.

Here is [bun's implementation of atob][4]. It's calling out to JSCore's native base64 decoder.

Basically every javascript environment has an absurdly optimized version of `atob`, except for node.js.

To say that I'm not allowed to make a trivial few-line change in order to make a function 70 times faster would be ridiculous. At that point I'm about ready to give up on node.js after my 14 years of using it, because it will be clear that node.js has adopted some very strange development philosophies: problems which Bun and Deno apparently do not have.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/atob
[2]: https://github.com/feross/buffer
[3]: https://github.com/denoland/deno/blob/v1.40.3/ext/web/lib.rs#L153
[4]: https://github.com/oven-sh/bun/blob/bun-v1.0.26/src/bun.js/bindings/ZigGlobalObject.cpp#L257