@anonrig 2024-02-06
I'm +1 on landing this and further optimize `atob`