@lemire 2024-02-06
(Before anyone gets angry at me... I stress that I did not oppose this PR nor do I plan to. There is no point in arguing with me about it, I am just trying to gain a better understanding. Just to make sure there is no misunderstanding I will also approve the PR after this comment.)


Here is what I think... The reason why we need to validate the string in JavaScript is because Buffer.from has effectively no error handling. It will just ignore any character it does not recognize.

```
> Buffer.from("aaéaa", 'base64').toString('latin1')
'i¦\x9A'
> Buffer.from("aaaa", 'base64').toString('latin1')
'i¦\x9A'
```

Yet the `Buffer.from` documentation says...

> A TypeError will be thrown if string is not a string or another type appropriate for Buffer.from() variants.


I think that if `Buffer.from` was whatwg compliant (at the C++ level) this PR would not be needed. If this, this means that this issue could be revisited later.

Please comment if you think I am wrong.