@lemire 2024-02-06
For context: https://github.com/nodejs/performance/issues/128