@lemire 2024-02-06
@chjj I am not offended, I just wanted to make sure that there was no disagreement.