@chjj 2024-02-07
[Here][1] is a [branchless version of the loop][2]. I don't know if everyone is okay with that, but it works well and gives a slight speedup over the current code in this PR.

[1]: https://github.com/chjj/node/commit/14c636c8c8093f84859f4665938271ccd199dbd6
[2]: https://github.com/chjj/node/blob/fix-atob2/lib/buffer.js#L1293