@lemire 2024-02-07
The new function will typically be faster when you anticipate that the result is valid, which should be often the case.