@tniessen 2024-02-07
Like most other collaborators, I believe I've never stood in the way of optimizing this function, so I don't appreciate the hostility in this discussion. Anyway, ...

> the `O(n*70)` worst case complexity of the loop isn't mentioned in the docs.

If you really want to throw O in here, be aware that `O(n*70) = O(n*100000) = O(n)`.

> This is a DoS vector to the unaware user and irresponsible IMO.

We rarely consider linear runtimes to be DoS vectors. If we did, one could argue that using a slow programming language interpreter or a slow CPU would be a DoS vector. There are likely many features in Node.js whose worst-case runtime far exceeds (any linear function of) the ideal runtime, yet we don't automatically consider those DoS vectors.

(Also, _actual_ security issues must be reported through HackerOne according to our security policy, and must not be disclosed through GitHub until they are resolved.)

> I really don't understand this logic. If it's never supposed to be used, why implement `atob` at all then? It's a liability to have it be __70 times__ slower than it needs to be. You might as well remove it at that point.

FWIW, I think most would agree with James and Anna that in an ideal world, nobody would be using this function, so optimization would be pointless, and in this non-ideal world, optimizing it encourages the use of an undesirable API. Still, the project's focus has shifted heavily towards interoperability and performance in recent years, and this PR doesn't add a new dependency like the other PR originally did, so it seems much more likely that the community will appreciate straightforward optimizations that don't add a ton of complexity.